chore: moving all functionality of the utils into languageClientManager
