+++ packages/salesforcedx-vscode-apex/test/vscode-integration/views/testOutline.test.ts
@@ -12,7 +12,8 @@ import * as fs from 'fs';
 import { createSandbox, SinonSandbox, SinonSpy, SinonStub } from 'sinon';
 import * as vscode from 'vscode';
 import { APEX_GROUP_RANGE, APEX_TESTS, FAIL_RESULT, PASS_RESULT } from '../../../src/constants';
-import { ClientStatus, LanguageClientUtils } from '../../../src/languageUtils/languageClientUtils';
+import { ClientStatus } from '../../../src/languageUtils/languageClientManager';
+import { LanguageClientUtils } from '../../../src/languageUtils/languageClientUtils';
 import { nls } from '../../../src/messages';
 import * as settings from '../../../src/settings';
 import { apexTestRunCacheService } from '../../../src/testRunCache';
